add skills handler and switch off of request method
